+++ src/domain/models/NeuronModel.kt
@@ -1,16 +1,14 @@
 package domain.models
 
-import java.math.BigDecimal
-
 data class NeuronModel(val inputSignalCount: Int) {
-    private val weights = mutableListOf<BigDecimal>()
+    private val weights = mutableListOf<Double>()
 
     init {
         for (i in 1 .. inputSignalCount)
-            weights.add(BigDecimal(Math.random()))
+            weights.add(Math.random())
     }
 
-    fun getWeights() : MutableList<BigDecimal> {
+    fun getWeights() : MutableList<Double> {
         return weights
     }
 }
